Refactors rest_app, adds launch_test_server.py
Refactors the rest_app into several smaller files each with related
API endpoints.

Adds launch_test_server.py to use for testing. This is a script that
starts a mysql DB, loads it with data, then starts the API.

The reason I used this solution is that the arxiv_tables.py didn't
work with SQLite and the fast-api test client didn't work for me. I
didn't see the value in altering the arxiv_tables to be compatible
with sqlite. I tried several ways to do a test client with fast-api
but they didn't work.